Use CSS to style Cite references
Output a default style of reference numbers and text for <ref>s and
<references>, use CSS content generation in supported browsers to render
per-language style.

Right now it provides pixel-perfect renditions of refs in enwiki and
eswiki, and initial support for fawiki (which can't be tested with current
phantomjs). The CSS is loaded from the new module 'ext.cite.style',
implemented in the Cite.php extension.

Further changes to Cite.php will be necessary to change it to use this
same CSS instead of system messages.

Test results (and two blacklisted results) changed due to new HTML
structure.

Related to bugs T51538, T45235 and T73803.

Bug: T86782
Change-Id: I21fbbd3247bf7801e5ef9bd5312f95777f4dd6ef