Hygiene: Main return as root structure (handle errors early)
Restructure code to return early for errors instead of nesting
conditionals. This leaves the outer tree of the function as
the natural flow of the primary purpose.

Makes code more resilient by ensuring that new code added to the
function will not execute under error conditions by removing the
need to keep everything encapsulated in the various levels of
error conditionals.

Change-Id: I1b4a67d344fd9843ca088d008487914f87b1c640