Skip registering RL modules depending on VisualEditor
When VisualEditor is not installed, there is no point in registered
resource loader modules that depends on it.  A use case is trying to run
tests for the MediaWiki tarball. It comes with Cite but without
VisualEditor.

The patch is based on GuidedTour patch by Matthew Flaschen
https://gerrit.wikimedia.org/r/#/c/305691/ for T143297

Remove the message "cite-ve-othergroup-item" which got introduced in
REL1_28 by 724d890 and is not in REL1_27.

Adjust base directory for REL1_27 since CiteHooks.php is at the root
when REL1_28 had it moved under includes/Cite.hooks.php

Change-Id: Idf769e0149f93c099a94b1b7a6cb203273dab881
(cherry picked from commit 5fcecf9e5f18a355327a883ecd745b9255f21173)