Make references not inherit bold or italic styling.
There is almost never a situation where this is desirable, yet it
happens quite a lot due to table headers etc having bold styling.
It confuses editors and tends to be less readable.

These style rules have been in en.wp:MediaWiki:Common.css for over 10
years or so, so probably a keeper.

Change-Id: If3d12383853a83d8ef14f1ec54c8c381b6c8f6a2