Make Cite pass phan-taint-check
Because of how arrays are handled, phan-taint-check thought all
return values from refArg() were escaped, where really only $dir
was. We also split the error method into the parse and noparse
case as separate functions so that phan can better analyse these calls.
In linkRef() we suppress the double escaping as the escaping used
is appropriate for inserting into wikitext.

Bug: T195009
Change-Id: I3e04c8cceae727e5470d4ae4fdb2404639f9bf33