Track pageviews only on content page views, not edits
This code is typically not executed on special pages and such, so we
assume the isArticle check doesn't make much of a difference. The main
difference we are aware of is that this will exclude previews.

Bug: T214493
Change-Id: I5155329b8a549adbd3b17c1f1014bb8bbc6768f4
(cherry picked from commit 4619cb24f65f420c2e6eca71557628a6c5e4ad2c)